fio-pod.sh updated with part responsible doing
register-tool-set --remote=ip_of_machine

with this change, it will not run register-tool-set against machines where it
was already ran

+++ fio-pod/fio-pod.sh
@@ -3,20 +3,24 @@
 # fio bench test for openshift pods
 # requires pbench_fio from https://github.com/distributed-system-analysis/pbench
 # to be installed on OSE master and OSE nodes
+# add 1
+
 
 # variables
 
 fiobin="/opt/pbench-agent/bench-scripts/pbench_fio"
-podfile="podfile"
+filepod="podfile"
 config="fio_testing"
 testtypes="read,write,rw,randread,randwrite,randrw"
 otheropt=""
 
+
+
 usage() {
-    printf -- "./fio-pod.sh -f filepod -c config -t testtype -o "other pbench_fio options" \n"
+    printf -- "./fio-pod.sh -f filepod -c config -d testdir\n"
     printf -- "where is:\n"
-    printf -- "podfile - list of pod's ip addresses where test will be run\n"
-    printf -- "config - string describing fio test config , eg fio-ceph-run1, fio-rhs-run1...etc\n"
+    printf -- "filepod - list of pod's ip addresses where test will be run\n"
+    printf -- "config - string describing test, eg fio-ceph-run1, fio-rhs-run1...etc\n"
     printf -- "testtype - fio test type to run, list comma separated test, eg: read,write, randread \n"
     printf -- "otheropt - run pbench_fio -h for list, quoted list of desired pbench_fio options - they will be passed to pbench_fio\n"
     exit 0
@@ -28,7 +32,7 @@ if [ "$EUID" -ne 0 ]; then
     exit 0
 fi
 
-opts=$(getopt -q -o f:c:t:o:h --longoptions "filepod:,fconfig:,testtypes:,otheropt:,help" -n "getopt.sh" -- "$@");
+opts=$(getopt -q -o f:c:t:o:h --longoptions "filepod:,config:,testtypes:,otheropt:,help" -n "getopt.sh" -- "$@");
 eval set -- "$opts";
 echo "processing options"
 while true; do
@@ -40,10 +44,10 @@ while true; do
 		        shift;
             fi
         ;;
-        -c|--fconfig)
+        -c|--config)
             shift;
             if [ -n "$1" ] ; then
-                fconfig="$1"
+                config="$1"
                 shift;
             fi
         ;;
@@ -54,7 +58,6 @@ while true; do
                 shift;
             fi
         ;;
-
         -o|--otheropt)
             shift;
             if [ -n "$1" ]; then
@@ -81,9 +84,9 @@ done
 # write pod file based on pods ip addresses - this will be run on OSE master
 write_podfile() {
     # just taking into considerations pods created by this script
-    rm $podfile.txt
+    rm $filepod.txt
     for pod in $(oc get pods | awk '{print $1}'| egrep -v 'NAME|docker-registry|router');do
-        oc describe pod $pod | grep IP | awk '{print $2}'>> $podfile.txt
+        oc describe pod $pod | grep IP | awk '{print $2}'>> $filepod.txt
     done
 }
 
@@ -93,15 +96,20 @@ write_podfile() {
 tool-set-register() {
     for node in $(oc get nodes | awk '{print $1}'  | grep -v NAME); do
         # requires SSH passwordless login -- configure this in advance
-        ssh -o StrictHostKeyChecking=no -nTx root@$node yum -y install pbench-*; source /opt/pbench-agent/profile; source /opt/pbench-agent/base
-        register-tool-set --remote=$node
-        printf "Tools registered for node : $node\n"
+        if ( ssh -o StrictHostKeyChecking=no -nTx root@$node '[ ! -f /var/lib/pbench-agent/tools.default ]' ) ; then
+            ssh -o StrictHostKeyChecking=no -nTx root@$node yum -y install pbench-*; source /opt/pbench-agent/profile; source /opt/pbench-agent/base
+            register-tool-set --remote=$node
+            printf "Tools registered for node : $node\n"
+        else
+            # here I assume if /var/lib/pbench-agent/tools.default exist then pbench packages are installed and register-tool-set already ran
+            printf "Tools are already registered for node : $node\n"
+        fi
     done
 }
 
 # main
 run_test() {
-        $fiobin --test-types="$testtypes"  --config=$fconfig --clients=$(cat $podfile.txt | awk -vORS=, '{ print $1 }' | sed 's/,$/\n/') $otheropt
+        $fiobin --test-types="$testtypes" --config=$config --clients=$(cat $filepod.txt | awk -vORS=, '{ print $1 }' | sed 's/,$/\n/') $otheropt
 }
 
 write_podfile
